datalake: unsupported json schema feature is a warn log
Logging with error level is a good default because in general we should
be able to build an intermediate representation (ir) from all the
schemas that were also accepted by the schema registry.

However, we intentionally do not support all json schemas, at least for
now. For schemas that we don't support we throw a nice exceptions which
helps differentiate unsupported schema vs unexpected schema.

To avoid error log noise and bad log lines in tests we log a warning for
such schemas instead of errors.