c/cloudchecks: increase default timeout to 10s
Redpanda uses a default 5s timeout for DNS lookups, so if the DNS lookup
for a cloud check operations times out or just takes ~5s then the self
check would fail.

We should be able to gracefully retry on at least 1 DNS lookup, so
increase the self check timeout to 10s to give the operations more time
to complete.

This was prompted by some occasional (~1/100) failures observed in Azure
CDT testing.