schema_registry: Prefer fragmented_vector
There are many uses of `std::vector<security::acl_binding>`, which
primarily stems from their use in:
* `cluster::create_acls_cmd_data::bindings`
* `cluster::delete_acls_result::bindings`

And is propagated through many APIs, which limits the number of
creations/deletions to about 1000 before an oversize alloc.

Regardless, this PR switches to fragmented_vector for associated
uses where possible, to avoid exacerbating the problem.

Signed-off-by: Ben Pope <ben@redpanda.com>