schema_registry: allow enabling import mode
Don't reject IMPORT mode when parsing a request, instead, allow enabling
it.

The preconditions to enable import mode are:
- globally: the schema registry must be empty
- subject-level: the subject must be empty

Soft-deleted schema versions are also considered. The reference
implementation has a feature where it allows moving to import mode even
if there are soft-deleted schemas, and it hard-deletes these schemas
while moving into import mode. We can consider implementing this later
on (left as a TODO).