tiered_storage_model_test: try to fix timeout
Test seems to timeout when shutting down background tasks.

One theory is that it's the log `tail -f` task that is hanging:
Redpanda has been stopped so will not product any more locks, so the
ssh_capture will never return with a new line and so we never check
the stop flag.

Solve this by stopping the background tasks before we shutdown the
Redpanda cluster.

CORE-12908