raft/basic_raft_fixture_test: deflake & hb fix
test_delayed_snapshot_request previously flaked because a cluster of a
bug in how it performed reconfiguration. the raft group would
reconfigure from 3 to 1 node, and then back from 1 to 3, but this test
used revision_id = 0 for all stages of this process.

This caused a race on in_flight_appends within follower metrics, because
if consensus uses an raii guard to increment and decrement
in_flight_appends to suppress heartbeats. If a guard opened, the group
reconfigured down, and then the group reconfigured up, and then the
guard expired, the guard would attempt to decrement a freshly created
follower metrics object, which would attempt to decrement the counter
below 0.