test: add grouping to the storage service
The purpose of the group is to support the semantics of object prefixing
and prefix listing in services like S3 without having to change the
model to actually implement this with explicit prefixing. In later
updates we'll add interfaces to examining the contents of the groups
which will model prefix listing.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>