tests/kgo: handle errors when polling for verifier status
It sometimes happens that the verifier status request returns an error
due to transient network issue or application being slow to start. In
this case the test failed completely as the producer status wasn't
reported even if the producer finished successfully. Added error
handling to the status loop to prevent those errors from failing the
tests.

Signed-off-by: Michał Maślanka <michal@redpanda.com>