tests/docker: run with azurite fork which does not leak conns
Temporary image built by plat eng team from
https://github.com/nvartolomei/Azurite/commit/c3b8e4f012fc77002f723ca40390752b1fb48ccb

---

In azurite all operations are put in an append queue (max 50 concurrent
ops) or a read queue (max 100 concurrent ops).

There are couple of issues with these queues but the biggest one is that
if ops are executed after client disconnects we try to attach event
handlers on a readablestream (body) which has been closed long ago. No
events are triggered for them any more (data/close/error) so we never
finish processing them and we hang there.

The fix checks if the stream is readable (alive) before installing event
handlers on it to make sure that these event handlers will get
triggered.

I'm not confident in the fix but pretty confident in the bug :)