ct/l1: add metastore frontend boilerplate
This adds the boilerplate required for the metastore frontend, that
routes RPCs to the metastore partition leader's shard. This consists
generally of three methods per underlying metastore call:
- do_FOO: performs the actual FOO operation on the local shard.
- FOO_locally: routes the request to the appropriate shard for the given
  FOO request and calls do_FOO on that shard.
- FOO: determines whether the target partition for the request has a
  leader that is local. If so, calls FOO_locally. Otherwise dispatches
  an FOO RPC to the leader.

This pattern, along with the concepts and templatized code for handling
routing, is copied almost verbatim from datalake/coordinator/frontend.*,
with the caveat that the concepts to verify request types are changed to
use topic_id_partition instead of topic name.

do_FOO methods are implemented as stubs, to be filled out in subsequent
commits.