dt/schema_registry: move test to mode-mutable suite
The `test_imported_schemas_with_dependencies_issues` test requires
posting schemas with specific versions/IDs, which won't work in
readwrite mode after an upcoming commit. Moving to the mode-mutable
suite allows enabling import mode for this test.

This is a straight copy/paste with no changes to the test logic.