`storage`: rework `compaction_backlog()` logic
The previous logic for `disk_log_impl::compaction_backlog()` was quite involved
and was mostly predicated on the fact that adjacent `segment` compaction was
the only form of compaction in `redpanda` at the time.

To estimate the number of bytes that must be read to fully compact the `log`,
search for the last dirty `segment` in the `log`, and sum over all of the bytes
that proceed it. This is an over-estimation of the number of bytes that
may be processed during compaction, as it does not take into account which
keys may be present or not in the dirty/clean `segment`s (ergo which `segment`s
will actually be read/rewritten by the built key-offset map during a round
of compaction), but this is fine for our purposes.

The `compaction_backlog()` is estimated as 0 if the `log` is not currently
eligible for compaction (i.e due to a dirty ratio below `min.cleanable.dirty.ratio`
and without a dirty timestamp exceeding `max.compaction.lag.ms`).