bazel: bump all golang deps
We want to keep all dependencies between the different go.mod files on
the same version, because in bazel we use a single dependency across all
modules. Docker is now resolving to a newer version, but
`github.com/moby/sys/user` was still on an old version, causing errors. I
would expect that `github.com/moby/moby` would have a dependency on the
newer version and that drag all the versions up but :shrug: