`redpanda`: pass disk availability to `controller_config()` functions
This also has the functional change of using availability instead of capacity
as the factor by which the normalization factor and setpoint is calculated
for the `compaction_controller_config`.

This makes more logical sense, since we should only be concerned with disk
space available to `redpanda` as an application instead of the disk's total
capacity when trying to make informed decisions on space management internally.