rpk/registry: create schema with id and version
This adds two optional flags to the `rpk registry schema create`
command, which allow users to specific a version and id to register the
schema with.

The id can only be set, and is required to be set, in import mode. This
is validated server side.

The version may only be set if the id is set. This is not technically
required server-side, but the franz-go implementation only passes on the
version if the id is set, so it makes sense to validate this inside rpk.

It would be possible to specify the version without the id in readwrite
mode, but in this case the server-side validates that the specified
version must be exactly the version that would be projected if it would
not have been specified. Therefore, there is no need to expose this
"functionality" to the user, and we can just require the id to always be
set if the version is specified.