dt/pb: add connectrpc to ducktape
After this commit we can now autogenerate admin clients for ducktape
tests.

These tests uncovered that we set the Content-Type header incorrectly
for protobuf responses, because the handler we're using the in the admin
server override the content type to be application/json unconditionally.

To support protobuf, we use a custom handler because we already set the
content type in our handler class.