ct: Add two-step shutdown to pipe
Pipeline components are supposed to use abort source from the underlying
pipeline. The idea is that the pipeline is a source of the data for
components and it also controls when they should stop. This allows us to
avoid passing abort source into many different places which share the
pipeline.

The problem is that it doesn't work well with
'ssx::sharded_service_container'. If the pipeline and all components are
services the container expects them to stop in certain order and
naturally, the pipeline is stopped last (because it's created first).
This makes 'stop' method of the pipeline component to get stuck.

This commit adds a 'shutdown' method. The methods triggers shutdown for
all components which use pipeline's abort source. The second step is a
'stop' method that works as before. The 'shutdown' call is optional.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>