ct/l1: plumb terms to metastore add_objects
Extends the add_objects metastore interface to require terms. This
commit mostly focuses on plumbing, passing terms from the metastore down
to the underlying state, for both the simple_metastore and replicated
metastore. Tests are updated to construct appropriate terms for test
extents.

Still, no validation is performed on these terms -- that will come in a
subsequent commit.