pbgen: support field mask merging
This extends the field mask to be easier to work with by making a path a
vector instead of a raw string. We then can traverse a path in codegen
instead of doing string manipulation. As part of this we enforce limits
on the size of field masks, since they are now stored as vectors.

Then we implement methods for validating and applying field masks to
merge two protos together. Note we follow the minimum amount of the
spec required here for field masks. Fancier operations like wildcards
or individual map updates are not supported.

Lastly we codegen the actual field mask operations themselves.