cl: Push API default behavior into model
Protobuf uses "0" as convention to indicate default for scalar values.
This convention is pushed into the model by taking those scalar values
and making them std::optional's.  When unset, cluster_link will use a
constant default value.  For the API, this would turn into a "0".  This
ensures consistency for users who wish to reconcile the settings.

Signed-off-by: Michael Boquard <michael@redpanda.com>