bazel: enhance tools shim mechanism
We have some llvm scripts in /tools which are wrappers around
`bazel run`. Specifically, they invoke the llvm tool of the same name in
the current toolchain. This fails if the cwd isn't under the workspace
root, so it would fail when invoked from outside the root, which
includes also the case of running any binary from bazel which then
invokes these wrappers (since binaries run in the runfiles dir which
is outside the workspace root).

This specific case occurs when ASAN or UBSAN failures occur, since
llvm-symbolizer is invoked, but the problem is a general one.

This change introduces a general trampoline in tools/bazel-tools/support
which knows how to:

(1) Change the cwd to be under the workspace root before invoking bazel
(2) Ensure that the target binary runs with the cwd set to the user cwd
   when the original shim was invoked

Item (2) requires the user of `--run_under=` targeting run.sh which is
yet another shim which knows how to determine the original cwd and cd
there before finally running the target.

Narrow the bazel-* ignore pattern to target only the top level.

Fixes CORE-13190.