rptest: add ASAN, UBSAN backtrace decoding tests
Check that when ASAN and UBSAN failures occur, the backtraces are
decoded immediately in the log. We check this by triggering a failure
of the required type then searching the logs for a known function name
in the backtrace.

This should prevent regressions like those seen in the past in this
area.