data_migrations_api_test: don't spam console
This test logged expected exceptions at ERROR level which go to the
console locally and in BK, causing a large amount of info and
preventing log streaming in BK.

Use exc_info=True to log the full exception at INFO level instead.

This reduces the output from 1 invocation of the test from ~6500 lines
to about 70, and there are more than 50 invocations due to
parametrization.