fixture.h: split into .h and .cc
redpanda/tests/fixture.h is a monster, try to tame it a bit by splitting
all the implementation out into a .cc, and clean up a lot of the headers
in the .h which are used only by the impl. Should improve LSP
responsiveness and compile times for fixture tests though they are
still relatively slow as fixture.h still includes application.h.

The header cleanup is incomplete due to a large number of undeclared
transitive dependencies.