cl/group_mirroring: use replica hwm if replica is collocated with task
If task is running on the same node that host the partition replica that
the committed offset is related to we can leverage that fact and query
for the high watermark. There is a high chance that the high watermark
coming from replica directly is more up to date than the one from health
report.

Signed-off-by: Michał Maślanka <michal@redpanda.com>