`storage`: use validated `max_ts` for `retention_ms`
The motivating case for `broker_time_based_retention` was the
fact that records with bad timestamps produced in the future
could lead to time-based retention being stuck indefinitely [1].

However, using the `broker_ts` can lead to unexpected
behavior when e.g. replicating data from an existing cluster
using MM2, as the timestamps of the Kafka records themselves are
correctly preserved, but internally, `redpanda` data structures
are not.

To avoid the potentially curious behavior of a divergence in
retention enforcement, go back to using the `max_timestamp`
for batches whose timestamps have been validated and unconditionally
set in the produce path (see: `v/kafka/server/handlers/produce_validation.cc`)
as of `v25.3.1`.

[1]:
* https://github.com/redpanda-data/redpanda/issues/9820
* https://github.com/redpanda-data/redpanda/pull/12991