kc/broker: fixed protocol version negotiation
Previously when broker sent an ApiVersionResponse, Redpanda kafka client
always used the request version to decode the response. This behavior is
incorrect.

According to KIP-511 when a broker (starting from version 2.4.0) doesn't
support the ApiVersionRequest that the client sent it will response with
ApiVersionsResponse version 0.

Changed the logic to peek for the error code first and then decode the
response with the correct version.

Additionally changed the way client software identifier and version are
built as they must match a regex hardcoded in Kafka broker.

Signed-off-by: Michał Maślanka <michal@redpanda.com>