ct: Downgrade the assertion
With the rm_stm in the write path the the frontend may see the
replicated batch with term set to -1. This happens in case of
idempotency viloation. The rm_stm uses the request from the list of
finished requests and this request could potentially come from the
snapshot which doesn't store term or from the log replay in which case
the term is not propagated to the request.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>