cl/group: fixed race condition leading to cg mirroring not being started
Previously the consumer group mirroring task relied on the partitions
leader table to check if current node is a partition leader and decide
if the task should run on the shard. This approach is prone to the race
condition as the notification may trigger task reconciliation first and
only then update partition leaders table.

In order to fix the race condition the method checking if the consumer
group mirroring should start was changed to check the partition
directly. This guarantees that the correct state is observed when
checking if task should run as the notification is fired after raft
protocol local state update.

Signed-off-by: Michał Maślanka <michal@redpanda.com>