`compaction`: move enhanced key functionality and fix `reducer_test`
The use of `compaction::enhance_key()` should be limited to local storage
compaction for indexing of a compaction map & de-duplication since the data
present in existing `.compaction_index`s is of this format.

However, this functionality has no real benefit for new compaction
implementations, since the extra bits encoded by `enhance_key()` are
not going to be used anywhere (and aren't even used in local storage anymore)
and are ultimately wasted space.

Duplicate the code in `compaction::is_latest_record_for_key()` back into a
function within an anonymous namespace in `storage/segment_deduplication_utils.cc`,
and remove the call to `enhance_key()` within the `compaction` namespace.

This code is left duplicated to avoid an extra `co_await` call to a generic
`compaction::` function in the hot path.

Also, fix the `SimpleReducer` test, which was subject to a bug in which
the non-enhanced key was being indexed but compared against the enhanced key
during de-duplication- because of the number of records produced with the
`linear_int_kv_batch_generator()`, this test was passing by accident.