cl/rpc: shadow_topic_report implementation
Shadow topic report aggregates the topic report from all the leaders of
the partitions of the topic. The implementation fans out RPCs to all
nodes hosting replicas of the topic. Each broker aggregates the report
from all shards and only responds if they think they are the leader
along with the minimum link revision applied across all shards of the
broker.