Add effective config output to rpk profile print
When printing a profile with `rpk profile print`, the parsed
configuration from disk is printed to the screen. However, we have
the ability to override these values with mechanisms like environment
variables.

We removed a previous implementation of this because it was confusing to
some users (see:
https://github.com/redpanda-data/redpanda/commit/49500411790750651f618e41026ad0da69805fe0).

This commit brings the feature back with some added provisions:

- The file-backed config is still shown by default
- The `-v/--verbose` shows the file-backed config as well as the
  effective config, with separate labels

Also, when using `rpk profile print --verbose`, we display extra info, which
may be confusing in a situation where we want to save and load a
profile via `rpk profile create`. This picks out the file-based
(rather than the effective) config when both are present.

Co-authored-by: Kat Batuigas <36839689+kbatuigas@users.noreply.github.com>
Co-authored-by: Rogger Vasquez <59714880+r-vasquez@users.noreply.github.com>