ct: recompute batch crc in L0 data plane
When we return batches to consumer, we take a materialized batch
(header, payload) from cloud storage, fiddling with the header so that
it has the correct base_offset, then recompute the CRC. Next, we swap
out the header in the materialized batch for the header from the
placeholder, and we retain the CRC from the materialized batch. The
apparent assumption here is that the fields under the kafka crc are
identical. Indeed, they nearly are identical, but nearly isn't good
enough. The difference I see in my experiments are that they differ only
by their attributes: the placeholder batch has CreateTime and the
materialized batch has LogAppendTime. Everything else is the same.

But that is still enough to fail the CRC check on the client side.
This patch fixes up all CRCs on created batches. A follow up to this PR
is to make the chain of custody and source of truth for metadata much
clearer.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>