cloud_topics/reconciler: move partition notifications into app
Use the manager to notify which partitions are eligible for
reconcilation, so we can move out a lot of the hard dependencies like
the partition manager and the metadata cache.

This simplifies the reconciler nicely, and we extract all the
notification wiring into a dedicated method as to not clutter up the
start method.