ct/l1: erase partition when removing last object leaves it empty
We were previously leaving behind empty partitions in the builder. This
meant that after a bad object is removed from the builder, the call to
replicated_metastore::add_objects() would examine the empty set of
objects meant for a given metastore partition, and expect there to be
terms routed for that partition. This resulted in the following error:

ERROR 2025-09-24 06:14:38,875 [shard 0:main] cloud_topics - replicated_metastore.cc:320 - No term metadata routed to partition 1