DisablingPartitionsTest::test_disable: deflake hwm
Per description in describe_topic: "this will omit any partitions which
do not have full metadata in the response: this means that if we are
unlucky and a partition returns NOT_LEADER due to a leadership
transfer... it will be missing"

This test asserts on all expected metadata being found.

The solution is to retry in cases where the expected partition metadata
is not found.