ct/ctp_stm_api: put ss::abort_source in APIs
The RTC node stuff is awkward to put in a constructor of a class, it's
much more straightforward to add to the methods, now we can properly
plumb around an abort source for this class, this removes some places
that were a little messy before.

As part of this we also simplify `reconciler::source::make_reader`
because we want it to take a required abort source, not an optional one
like the normal reader config.