rptest: adjust type-check.sh for pre-commit hooks
Remove `-i` from docker run, since no stdin is needed.

This is to allow running the check in a pre-commit hook, which would
otherwise error with:
```
the input device is not a TTY
```

This is safe because the type checker needs no input from stdin. Signals
such as ctrl+C are still passed through correctly with docker.