cluster: wcr support for uuid override
WCR subsystem uploads cluster manifests periodically to object storage
keyed by cluster UUID and a sequence number.

When a cluster is re-created, it gets a new cluster UUID and the
sequence number is initialized as the highest seen so far in the bucket.

Here is an example of a cluster (assigned <uuid-a> initially) after it
uploaded 3 manifests then was re-created (assigned <uuid-b> this time).

```
/
+- cluster_metadata/
   +- <uuid-a>/manifests/
   |  +- 0/cluster_manifest.json
   |  +- 1/cluster_manifest.json
   |  +- 2/cluster_manifest.json (topic A exists)
   + <uuid-b>/manifests/
      +- 3/cluster_manifest.json (empty)
      +- 4/cluster_manifest.json (empty)
```

Attempting WCR searches for highest sequence number ignoring current
cluster uuid (i.e. <uuid-b). If we run WCR, we will pick
`/cluster_metadata/<uuid-a>/manifests/2/cluster_manifest.json` as source
of truth.

This eventually would lead to creation of another manfiest:

```
/
+- cluster_metadata/
   # ...
   +- <uuid-b>/manifests/
      # ...
      +- 5/cluster_manifest.json (topic A recovered)
```

This algorithm (A1) is known to reach an undesirable result in the
following scenario: Imagine that we start with the same state as above
prior to restore but we do not run restore but instead re-create the
cluster once more. This can happen either because an operator wanted to
restart recovery process from scratch or because a disaster happened in
short span before recovery had a chance to finish/start. And we end with
this picture:

```
/
+- cluster_metadata/
   +- <uuid-a>/manifests/
   |  +- 0/cluster_manifest.json
   |  +- 1/cluster_manifest.json
   |  +- 2/cluster_manifest.json (topic A exists)
   + <uuid-b>/manifests/
   |  +- 3/cluster_manifest.json (empty)
   |  +- 4/cluster_manifest.json (empty)
   + <uuid-c>/manifests/
      +- 5/cluster_manifest.json (empty)
      +- 6/cluster_manifest.json (empty)
```

If we run the same algorithm (A1) then it will pick
`/cluster_metadata/<uuid-b>/manifests/4/cluster_manifest.json`. However,
we want to recover from <uuid-a> because that's where our data lives.

To support this, we add an optional override to cluster recovery
subsystem where user can explicitly choose from which cluster UUID to
restore.

---

This also works with `cloud_storage_cluster_name` and has guardrails to
prevent restoring from a UUID belonging to a different cluster and has
guardrails to prevent restoring from a UUID belonging to a different
cluster.