ct: rename underlying to ctp
Using a name like underlying is pretty common for type aliases (e.g.
underlying = std::map), but in the context of the L0 reader the backing
partition is a cloud topic partition and it is less confusing to simply
call the member by its real name instead of a generic name.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>