raft: move follower_index_metadata from types.h to follower_stats.h
For coordinated compaction we need to add fields
to follower_index_metadata. The type of one of these fields depends on
the class to be added, which depends on an RPC type defined in types.h.

This would introduce a circular header dependency
if follower_index_metadata remained in types.h.
Move it to follower_stats.h