config: turn auto enable enterprise feature on by default
Change the value of core_balancing_continuous and
partition_autobalancing_mode to 'on' by default. This behavior should be
enabled from v25.3 onwards.

Changes in the test:
With autoenabled enterprise features, every cluster without a license that doesn't
explicitly turn them all off will be in license violation. Thus the part
of the tests that checks for that has been removed.