cluster: add cloud topic tombstone to topic_table
Adds tombstones to the topic_table that is created for cloud_topics when
they are deleted (if remote_delete is specified). The mechanics here are
pretty much identical to what's used for Iceberg and tiered storage.

We'll expect some fiber to examine the topic_table for cloud topic
tombstones, delete the topic state in the L1 metastore, and then remove
the tombstone by calling topics_frontend::purged_topic().