metastore: add min offset lower bound to timequeries
Listing offsets should give a strongly consistent response in face of
operations like DeleteRecords or retention. Since we might be delayed in
pushing the new start offset to L1, we add a parameter to timequeries to
provide a lower bound. This allows us to prevent additional L1 IO for
timequeries.