tuner: Add banners to tuner start log lines
The net tuner logging is very verbose as every sub-tuner reruns the same
basic functions again.

This leads to lots of output which can make it hard to find a single sub
tuner in the output.

Add some banner to the initial log lines to make it easier for the eye
to find the start of the sub-tuner.