k/server: track group_id per-connection
Track the `group_id` on relevant kafka APIs.

Note that the OffsetDelete API is excluded, as it is considered to be a
kafka admin API and would not identify a consumer group client.

Similarly, the LeaveGroup API is also excluded, as it can be used a
kafka admin API command.

The OffsetFetch API is included if the fetch is for a single group,
because consumer clients typically start with fetching the latest
offsets for their group.