storage/log: calculate cleanly_compacted_prefix_offset
This PR is part of coordinated tombstone removal project.

This change is to collect the offset that binds the longest known
cleanly compacted log prefix, i.e. the log prefix where each key is
present there no more than once. This offset is called MCCO in the RFC.

We cache MCCO known from previous queries, bump it to log start in case
it was prefix-truncated and check further segments one by one until we
find one that is not cleanly compacted or get to the end of the log.
The beginning of such segment or the end of the log will be the new MCCO

Complexity is amortized linear, as it caches the result so segments are
not re-scanned.