{storage,compaction}/types: add max tombstone removal offset to config
Coordinated compaction mechanism is going to restrict tombstone removal
on individual nodes based on compaction states of all replicas. Max
offset to remove tombstones will be calculated in raft layer.

This change is to pass it to storage layer via compaction_config. For
now init it with the same value as max compaction offset. This matches
current behaviour, where tombstone removal is not restricted
specifically, but is performed in the same offset range as normal
compaction. It'll be overridden later using storage::stm_manager.