cloud_io: Use bulk delete for Linode Object Storage
Linode has very low rate limits for DELETE operations (100 req/s but it
depends on endpoint type). Delete operations issued by Redpanda can be
quite spiky especially when DeleteRecords is used or retention is
updated. Together with the fact that Tiered Storage retention in
Redpanda is non-incremental it is probable for retention mechanism to
stop making progress.

Use bulk operations to improve the situation.

This fixes https://redpandadata.atlassian.net/browse/CORE-13400.

A more comprehensive solution is tracked in issue 13399.