ct/l1: Add retry helper for the metastore
This adds a templated retry helper for metastore operations and replaces
the existing retry logic in add_objects_with_retry with the helper. It
also uses the helper to add retries to metastore builder creation and
metastore operations in the L1 reader.

There are existing reconciler tests that retries work for add objects
and which serve as a test of the retry helper.