admin: datalake protobuf spec for get_state
Adds protobuf definitions for getting the datalake coordinator state as
an admin endpoint. This will be an important endpoint for upcoming
upgrades, where some breaking changes will occur to Iceberg tables.

Existing Iceberg Topics will need to:
1. pause translation
2. quiesce pending state before upgrading
3. perform some manual data rewriting
4. upgrade
5. reenable Iceberg with the upgraded schema mapping.

This endpoint in particular will be used in step 2.

I considered adding a more specialized endpoint to check if there is any
pending state (i.e. return a simple boolean), but thought the visibility
of seeing all the topic state would be useful in practice.