ducktape: Speedup metric gathering on shutdown
When running a basic DT self test we see that "shutdown" is fairly slow:

```
[INFO  - 2025-10-24 14:09:12,672 - redpanda - stop - lineno:4487]: RedpandaService-0-138017904872800: stopping service
[DEBUG - 2025-10-24 14:09:15,199 - remoteaccount - _log - lineno:186]: root@redpanda-0: Running ssh command: pgrep --list-full --exact redpanda
```

On Azure in `RedpandaServiceSelfTest.test_start` we see a total run time of:

```
run time:         6.869 seconds
```

Looking into it we see that this is because when scraping metrics we query the
metrics endpoint N times. This is quite slow as each attempt takes in the
region of XXX milliseconds.

Improving the metrics scraping code to only scrape once we get the following
improvement:

```
[INFO  - 2025-10-24 14:07:23,215 - redpanda - stop - lineno:4487]: RedpandaService-0-135056967880080: stopping service
[DEBUG - 2025-10-24 14:07:23,654 - remoteaccount - _log - lineno:186]: root@redpanda-0: Running ssh command: pgrep --list-full --exact redpanda
```

```
run time:         4.775 seconds
```